table_service/requirements.txt
asgiref==3.9.0
Django==5.2.4
django-bootstrap5==25.1
django-cors-headers==4.7.0
django-filter==25.1
django-tables2==2.7.5
djangorestframework==3.16.1
Markdown==3.8.2
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
tablib==3.8.0
